Move getClausesSplitOnStopWords file to freeAuxiliaryParticipleOrder folder
